backend/requirements.txt
Django==4.2.7
graphene-django==3.1.5
django-graphql-jwt==0.3.4
mysqlclient==2.2.0
python-dotenv==1.0.0
requests==2.31.0
django-cors-headers==4.3.1
gunicorn==21.2.0
whitenoise==6.7.0
reportlab==3.6.12
